These are combat script changes
They were changed. Just some more progress.

+++ FetchQuest/Assets/Scripts/Combat/Consumable.cs
@@ -5,7 +5,7 @@ public class Consumable : Item {
 
 
     public enum ConsumeType {healthPotion, staminaPotion, type3 } //determines what kind of an effect a consumable could have
-    int NumValue; //this indicates the 'power of a consumable item. 
+    double NumValue; //this indicates the 'power of a consumable item. 
     public ConsumeType type;
